db_dump.sql al 26 marzo 2020. Cambio de dirname a soa
