+++ soa/service/misc/common.py
