+++ soa/service/misc/helperstr.py
