+++ soa/service/requirements.txt
@@ -1,0 +1,4 @@
+flask-restplus
+gunicorn
+psycopg2
+Werkzeug==0.16.1
